Added Reverse Control Powerup
Unity Changes: Created prefab for powerup. Added prefab to powerup list.

Script Changes: Made changes to paddles script and powerup manager. Occupied switch case for Reverse Controls. The current player's controls are reversed. Created a method in paddles that reverses the controls. It is activated if a bool is true. When the method is called in PowerupManager, the paddles script recognizes the bool as true and reverses the controls.